bolo/pyproject.toml
[tool.poetry]
name = "bolo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.19.1"
Flask-HTTPAuth = "^4.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
